refactor: reorganize test directories — tests/unit/ and tests/playwright/
- Move vitest files from test/ to tests/unit/
- Move Playwright mock tests to tests/playwright/
- Update all import paths, configs, and CI references
