+++ packages/vscode/tests/playwright/assert-view.spec.ts
