+++ packages/vscode/tests/playwright/auto-close.spec.ts
