+++ packages/vscode/tests/playwright/codegen.spec.ts
