+++ packages/vscode/tests/playwright/decorations.spec.ts
