+++ packages/vscode/tests/playwright/global-errors.spec.ts
