+++ packages/vscode/tests/playwright/highlight-locators.spec.ts
