+++ packages/vscode/tests/playwright/list-files.spec.ts
