+++ packages/vscode/tests/playwright/pick-selector.spec.ts
