+++ packages/vscode/tests/playwright/pnp.spec.ts
