+++ packages/vscode/tests/playwright/project-setup.spec.ts
