+++ packages/vscode/tests/playwright/project-tree.spec.ts
