+++ packages/vscode/tests/playwright/settings.spec.ts
