+++ packages/vscode/tests/playwright/trace-viewer.spec.ts
