+++ packages/vscode/tests/playwright/utils.ts
@@ -16,7 +16,7 @@
 
 import { expect as baseExpect, test as baseTest, Browser, BrowserContextOptions, chromium, Page } from '@playwright/test';
 // @ts-ignore
-import { Extension } from '../dist/extension';
+import { Extension } from '../../dist/extension';
 import { TestController, VSCode, WorkspaceFolder, TestRun, TestItem } from './mock/vscode';
 
 import crypto from 'crypto';
@@ -111,7 +111,7 @@ export const expect = baseExpect.extend({
 let _l10Bundles: [bundle: string, contents: Record<string, string>][];
 function l10Bundles() {
   if (!_l10Bundles) {
-    const l10nDir = path.resolve(__dirname, '..', 'l10n');
+    const l10nDir = path.resolve(__dirname, '../..', 'l10n');
     _l10Bundles = fs.readdirSync(l10nDir).map(file => [file, JSON.parse(fs.readFileSync(path.join(l10nDir, file), 'utf-8'))]);
   }
   return _l10Bundles;
@@ -123,7 +123,7 @@ export const test = baseTest.extend<TestFixtures, WorkerOptions>({
   vsCodeVersion: [1.86, { option: true, scope: 'worker' }],
 
   vscode: async ({ browser, vsCodeVersion }, use) => {
-    const vscode = new VSCode(vsCodeVersion, path.resolve(__dirname, '..'), browser);
+    const vscode = new VSCode(vsCodeVersion, path.resolve(__dirname, '../..'), browser);
     await use(vscode);
 
     if (process.env.VALIDATE_L10N) {
