+++ packages/vscode/tests/unit/compiler.test.ts
