+++ packages/vscode/tests/unit/mode-detect.test.ts
