+++ packages/vscode/tests/unit/vscode-stub.ts
